% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate-commands.R
\name{add_to_array}
\alias{add_to_array}
\alias{remove_from_array}
\title{Add or Remove Subvariables to an Array Variable}
\usage{
add_to_array(x, ..., position = Inf)

remove_from_array(x, ...)
}
\arguments{
\item{x}{TKTKTK}

\item{...}{TKTKTK}

\item{position}{TKTKTK}
}
\description{
Modify an existing array variable by adding or removing subvariables.
Removing subvariables is a schema command and so must be run before any
non-schema commands are run.
Based on the Crunch Automation command
\href{https://help.crunch.io/hc/en-us/articles/360045427871}{ALTER ARRAY}
}
\seealso{
Other creating variables functions: 
\code{\link{categorical_array}()},
\code{\link{categorical_case_when}()},
\code{\link{categorical_cut}()},
\code{\link{categorical_interaction}()},
\code{\link{convert_to_text}()},
\code{\link{create-vars}},
\code{\link{multiple_dichotomy}()},
\code{\link{multiple_selection}()}
}
\concept{creating variables functions}
