% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-dictionary-code-generation.R
\name{apply_excel_dictionary}
\alias{apply_excel_dictionary}
\title{Generate a Crunch Automation Script from an excel variable editor}
\usage{
apply_excel_dictionary(dataset, excel_file, out_file = NULL, submit = TRUE)
}
\arguments{
\item{dataset}{A \code{CrunchDataset} the variable editor describes}

\item{excel_file}{Filepath to an .xlsx file, generally an excel document
created by \code{\link[=create_excel_dictionary]{create_excel_dictionary()}} that you've modified}

\item{out_file}{A filepath to save the Crunch Automation Script to (\code{NULL},
the default, places the file in your temporary directory)}

\item{submit}{Logical, indicating whether to submit the generated script to be run
(defaults to \code{TRUE})}
}
\value{

}
\description{
Details forthcoming. Experimental.
}
