% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-tibble.R
\name{as_cr_tibble}
\alias{as_cr_tibble}
\title{Flatten a Crunch Cube}
\usage{
as_cr_tibble(x, ...)
}
\arguments{
\item{x}{a CrunchCube}

\item{...}{further arguments passed on to \code{tibble::as_tibble()}}
}
\description{
Crunch Cubes can be expressed as a long data frame instead of a
multidimensional array. In this form each dimension of the cube is a variable
and the cube values are expressed as columns for each measure. This is useful
both to better understand what each entry of a cube represents, and to work
with the cube result using tidyverse tools.
}
\details{
The \code{cr_tibble} class is a subclass of \code{tibble} that has extra metadata
to allow \code{ggplot::autoplot()} to work. If you find that this extra
metadata is getting in the way, you can use \code{as_tibble()} to get
a true \code{tibble}.
}
