% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-tibble.R
\name{as_tibble.CrunchCube}
\alias{as_tibble.CrunchCube}
\title{Flatten a Crunch Cube}
\usage{
\method{as_tibble}{CrunchCube}(x, show_metadata = TRUE, ...)
}
\arguments{
\item{x}{a CrunchCube}

\item{show_metadata}{By default this function returns additional metatada
about the. In particular the unweighted counts for each entry, and whether
that entry represents a missing value. If false the metadata is suppressed.}

\item{...}{futher arguments passed on to \code{tibble::as_tibble()}}
}
\description{
Crunch Cubes can be expressed as a long data frame instead of a
multidimensional array. In this form each dimension of the cube is a variable
of the dataframe and the cube values are expressed columns for each measure.
This is useful both to better understand what each entry of a cube
represents, and to work with the cube result using tidyverse tools.
}
