% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{AutomationCrunchDataset-class}
\alias{AutomationCrunchDataset-class}
\alias{AutomationCrunchDataset}
\title{A Crunch Dataset With Crunch Automation Commands "At The Ready"}
\description{
This is a subclass of \code{crunch::CrunchDataset} that has a field for recording
Crunch Automation commands that will get run when \code{collect} or \code{calculate}
is called.
}
