% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{autoplot.CrunchCube}
\alias{autoplot.CrunchCube}
\title{Autoplot methods for Crunch Objects}
\usage{
\method{autoplot}{CrunchCube}(x, plot_type = c("dot", "grid", "bar"),
  measure = "count")
}
\arguments{
\item{x}{a CrunchCube, or CrunchVariable}

\item{plot_type}{One of \code{"dot"}, \code{"grid"}, or \code{"bar"} which indicates the plot family
you would like to use. Higher dimensional plots add color coding or facets depending
on the dimensionality of the data.}

\item{measure}{The measure you wish to plot. This will usually be \code{"count"}, the default
but can also be \code{".unweighted_counts"} or any other measure stored in the cube.}
}
\description{
Generates ggplot representations of CrunchVariables and CrunchCubes
}
\details{
The Crunch autoplot methods provide plots which are optimized for crunch objects
and allow you to plot them without bringing the full object into memory. The plots produced
are:
\itemize{
\item Numeric variables: A histogram
\item Categorical variables: a sorted bar or dot plot
\item 2d CrunchCubes:
}
}
