% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automation-text-helpers.R
\docType{data}
\name{ca}
\alias{ca}
\title{Crunch Automation Syntax Helpers}
\format{
An object of class \code{list} of length 13.
}
\usage{
ca
}
\description{
Functions that help create special keywords for use in Crunch Automation commands.
}
\details{
The \code{ca} object contains the following objects and functions:
\itemize{
\item \code{category(label = NA, code = NA, missing = FALSE, ...)}: A function for creating a
category label which has a string label, an integer code, and a logical indicating
whether the category is a missing category. Used within other functions that may use
other arguments in \code{...}.
\item \code{copy}: Returns \code{COPY}, used to indicate that metadata should be copied from source
in some circumstances.
\item \code{dots(x, y)}: A function that takes two aliases (\code{x}, \code{y}) and returns \code{x...y} which
is shorthand in Crunch Automation for selecting the variables between x & y.
\item \code{like(x)}: A function that takes a single string and wraps it in \code{LIKE("")}, which in
Crunch Automation interprets it similar to "SQL's" LIKE statement, where a
\verb{\%} matches any string and \verb{_} matches a single character.
\item \code{regex(x)}/\code{regexp(x)}: Functions that take a single string in wrap in \code{REGEX("")}/
\code{REGEXP("")}, which Crunch Automation interprets.
\item \code{use_descriptions}: Returns \verb{USE DESCRIPTIONS} which indicates the metadata should come
from the descriptions in some circumstances.
\item \code{use_titles}: Returns \verb{USE TITLES} which indicates the metadata should come
from the titles in some circumstances.
\item \code{keyword(x)}: Returns the text unquoted, useful if you want to use a Crunch Automation
keyword not yet supported by \code{crplyr}.
}
}
\examples{
\dontrun{
ds \%>\%
    mutate(convert_to_text(ca$like("\%other"), description = ca$copy))
}
}
\keyword{datasets}
