% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate-automation.R
\name{categorical_array}
\alias{categorical_array}
\title{Create a categorical array}
\usage{
categorical_array(
  ...,
  labels = NULL,
  title = NULL,
  description = NULL,
  notes = NULL
)
}
\arguments{
\item{...}{One or more crunch variables selected using their alias,
\code{\link[dplyr:across]{dplyr::across()}}, or the special selection functions in \code{\link{ca}}.}

\item{labels}{The labels for the subvariables (also called the subvariable
names in the crunch R package).}

\item{title}{Optional, the title of the variables created (also called the variable name
in the crunch R package)}

\item{description}{Optional, the description of the variables created}

\item{notes}{Optional, the notes of the variable created}
}
\description{
Create a categorical array from one or more existing variables.
Based on the Crunch Automation command
\href{https://help.crunch.io/hc/en-us/articles/360042039392}{CREATE CATEGORICAL ARRAY}.
}
\details{
If \code{...} is specified using aliases or \code{dplyr::across()}, then
the other arguments can take functions based on the variables
selected using \code{\link{crplyr-formula-notation}}.
}
\seealso{
Other creating variables functions: 
\code{\link{convert_to_text}()},
\code{\link{create-vars}}
}
\concept{creating variables functions}
