% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate-commands.R
\name{categorical_case_when}
\alias{categorical_case_when}
\title{Create a Categorical Variable by Assign Expressions to Categories & Variables}
\usage{
categorical_case_when(
  ...,
  cases = NULL,
  title = NULL,
  description = NULL,
  notes = NULL
)
}
\arguments{
\item{..., }{cases Expressions where the right hand side is a Crunch logical
expression (or \code{TRUE} to indicate that it should be met if all other
expressions were not met) and the right hand side is a category or
existing categorical variable.}

\item{title}{Optional, the title of the variables created (also called the variable name
in the crunch R package)}

\item{description}{Optional, the description of the variables created}

\item{notes}{Optional, the notes of the variable created}
}
\description{
Create a Categorical Variable by assigning expressions variables to either
categories directly or to the values of existing categorical variables.
The new variable's alias is based on the name of the argument within
\code{mutate()}. Based on the Crunch Automation command
\href{https://help.crunch.io/hc/en-us/articles/360042039192}{CREATE CATEGORICAL CASE} &
\href{https://help.crunch.io/hc/en-us/articles/360042457871}{CREATE CATEGORICAL CASE THEN VARIABLE}.
}
\seealso{
Other creating variables functions: 
\code{\link{categorical_array}()},
\code{\link{categorical_cut}()},
\code{\link{categorical_interaction}()},
\code{\link{convert_to_text}()},
\code{\link{create-vars}},
\code{\link{multiple_dichotomy}()},
\code{\link{multiple_selection}()}
}
\concept{creating variables functions}
