% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate-commands.R
\name{categorical_cut}
\alias{categorical_cut}
\title{Create a Categorical Variable by Cutting a Numeric Variable into Categories}
\usage{
categorical_cut(
  ...,
  breaks = NULL,
  closed = NULL,
  labels = NULL,
  missing = NULL,
  new_aliases = NULL,
  title = NULL,
  description = NULL,
  notes = NULL
)
}
\arguments{
\item{...}{One or more crunch variables selected using their alias,
\code{\link[dplyr:across]{dplyr::across()}}, or the special selection functions in \code{\link{ca}}.}

\item{breaks}{TKTKTK}

\item{closed}{TKTKTK}

\item{labels}{TKTKTK}

\item{missing}{TKTKTK}

\item{new_aliases}{If only one variable is specified, this argument is ignored,
the alias comes from the name of the argument in mutate (and if it's the same
the variable will be modified in place, while if different it will create a new
variable). If more than one variable is specified \code{NULL}, the default, will modify
the variables in place, or a vector of aliases for the new variables (or a
\code{\link{crplyr-formula-notation}} that creates one).}

\item{title}{Optional, the title of the variables created (also called the variable name
in the crunch R package)}

\item{description}{Optional, the description of the variables created}

\item{notes}{Optional, the notes of the variable created}
}
\description{
Create a Categorical Variable by cutting the numeric values into categories
where values are between the breakpoints provided by \code{breaks}. The new variable's
alias is based on the aliases in \code{new_aliases} by default, but if it is
\code{NULL} and there is only a single variable, then it is based on
the name of the argument within \code{mutate()}.
Based on the Crunch Automation command
\href{https://help.crunch.io/hc/en-us/articles/360042458431}{CREATE CATEGORICAL CUT}
}
\details{
If \code{...} is specified using aliases or \code{dplyr::across()}, then
the other arguments can take functions based on the variables
selected using \code{\link{crplyr-formula-notation}}.
}
\seealso{
Other creating variables functions: 
\code{\link{categorical_array}()},
\code{\link{categorical_case_when}()},
\code{\link{categorical_interaction}()},
\code{\link{convert_to_text}()},
\code{\link{create-vars}},
\code{\link{multiple_dichotomy}()},
\code{\link{multiple_selection}()}
}
\concept{creating variables functions}
