% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect.CrunchDataset}
\alias{collect.CrunchDataset}
\title{Collect a crunch dataset from the server}
\usage{
\method{collect}{CrunchDataset}(x, ...)
}
\arguments{
\item{x}{A crunch Dataset}

\item{...}{Not used, included for s3 method consistancy.}
}
\value{
a tibble
}
\description{
This function brings a crunch dataset into memory so that you can
work with the data using R functions. Since this can create a long running
query it is recommended to try to filter the dataset down as much as possible
before running \code{collect()}.
}
\examples{
\dontrun{
ds \%>\% 
   select(cyl, gear) \%>\% 
   collect()
}
}
