% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{collect}
\alias{collect}
\title{Collect a crunch dataset from the server}
\usage{
collect(ds)
}
\arguments{
\item{ds}{A crunch Dataset}
}
\value{
a tibble
}
\description{
This function brings a crunch dataset into memory so that you can
work with the data using R functions. Since this can create a long running
query it is recommended to try to filter the dataset down as much as possible
before running \code{collect()}.
}
\examples{
\dontrun{
ds \%>\% 
   select(cyl, gear) \%>\% 
   collect()
}
}
