% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automation-processing.R
\name{compute}
\alias{compute}
\alias{compute.AutomationCrunchDataset}
\title{Run Crunch Automation Query}
\usage{
\method{compute}{AutomationCrunchDataset}(x, name = NULL, ...)
}
\arguments{
\item{x}{An \code{AutomationCrunchDataset} (a \code{CrunchDataset} modified with \code{\link[=mutate]{mutate()}} or similar)}

\item{name}{ignored}

\item{...}{ignored}
}
\value{
invisibly, the CrunchDataset after running the command
}
\description{
Run the commands that have been created by crplyr's \code{\link[=mutate]{mutate()}} and
similar commands.
}
\examples{
\dontrun{
ds <- ds \%>\%
    mutate(
       taste = categorical_array(
           taste_a, taste_b, 
           labels = c("Brand A", "Brand B"), 
           title = "Taste Rating"
       )
    ) \%>\%
    compute()
}
}
\seealso{
Other automation script commands: 
\code{\link{mutate}()},
\code{\link{show_query}()}
}
\concept{automation script commands}
