% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate-commands.R
\name{convert_to_text}
\alias{convert_to_text}
\title{Convert a Variable's Type}
\usage{
convert_to_text(
  ...,
  new_aliases = NULL,
  title = NULL,
  description = NULL,
  notes = NULL
)
}
\arguments{
\item{...}{One or more crunch variables selected using their alias,
\code{\link[dplyr:across]{dplyr::across()}}, or the special selection functions in \code{\link{ca}}.}

\item{new_aliases}{If only one variable is specified, this argument is ignored,
the alias comes from the name of the argument in mutate (and if it's the same
the variable will be modified in place, while if different it will create a new
variable). If more than one variable is specified \code{NULL}, the default, will modify
the variables in place, or a vector of aliases for the new variables (or a
\code{\link{crplyr-formula-notation}} that creates one).}

\item{title}{Optional, the title of the variables created (also called the variable name
in the crunch R package)}

\item{description}{Optional, the description of the variables created}

\item{notes}{Optional, the notes of the variable created}
}
\description{
When called from \code{\link{mutate}}, converts the variables type either as a new variable, or in place
(in which case it is considered a "schema command" and must be run before any
"non-schema commands").
Based on the Crunch Automation commands
\href{https://help.crunch.io/hc/en-us/articles/360042247191-CONVERT-command}{CONVERT} and
\href{https://help.crunch.io/hc/en-us/articles/360047136371-CREATE-CONVERT-command}{CREATE CONVERT}.
}
\details{
If \code{...} is specified using aliases or \code{dplyr::across()}, then
the other arguments can take functions based on the variables
selected using \code{\link{crplyr-formula-notation}}.
}
\seealso{
Other creating variables functions: 
\code{\link{categorical_array}()},
\code{\link{create-vars}}
}
\concept{creating variables functions}
