% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate-automation.R
\name{create-vars}
\alias{create-vars}
\title{Functions for Creating Variables Inside \code{mutate()}}
\description{
When creating and modifying variables in a \code{CrunchDataset} with \code{mutate()},
they must be created using the functions listed below.
}
\seealso{
Other creating variables functions: 
\code{\link{categorical_array}()},
\code{\link{convert_to_text}()}
}
\concept{creating variables functions}
