% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-excel-dictionary.R
\name{create_excel_dictionary}
\alias{create_excel_dictionary}
\title{Save an Excel document that describes the variables in a Crunch Dataset (EXPERIMENTAL)}
\usage{
create_excel_dictionary(ds, out_file = NULL, overwrite = FALSE)
}
\arguments{
\item{ds}{A \code{CrunchDataset}}

\item{out_file}{File path to save the excel \code{.xlsx} document}

\item{overwrite}{Whether to overwrite the file (default \code{FALSE})}
}
\description{
Create an MS Excel document that describes the variables and their metadata
(type, alias, title, description, notes, subvariables and categories). This
function is under active development and so its structure may change.
}
