% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automation-processing.R
\name{crplyr-formula-notation}
\alias{crplyr-formula-notation}
\title{Use Argument Values Baed on Input Variables}
\description{
Many \code{crplyr} commands have arguments that allow the use of
formulas (using the compact lambda style functions using \code{~} or functions,
as in the \code{purrr} package), where the function will be given a \code{data.frame}
containing the selected Crunch variables to operate on.
}
\examples{
\dontrun{
ds <- ds \%>\%
    mutate(
       taste = categorical_array(
           taste_a, taste_b, 
           # Labels will be Title Case form of subvariable titles
           labels = ~stringr::str_to_title(title(.)), 
           title = "Taste Rating"
       )
    )
}
}
