% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate-automation.R
\name{crunch_var_df-meta}
\alias{crunch_var_df-meta}
\alias{aliases,crunch_var_df-method}
\alias{titles,crunch_var_df-method}
\alias{descriptions,crunch_var_df-method}
\alias{notes,crunch_var_df-method}
\title{Crunch variable metadata for \code{crunch_var_df} objects}
\usage{
\S4method{aliases}{crunch_var_df}(x)

\S4method{titles}{crunch_var_df}(x)

\S4method{descriptions}{crunch_var_df}(x)

\S4method{notes}{crunch_var_df}(x)
}
\arguments{
\item{x}{A \code{crunch_var_df} object which contains \code{CrunchVariable}
objects.}
}
\value{
A character vector
}
\description{
Get variable aliases, titles (also called "variable names" by the
crunch R package), descriptions and notes from the variables
collected into a data.frame when using \url{crplyr-formula-notation}.
}
