% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_.CrunchDataset}
\alias{filter_.CrunchDataset}
\title{Filter a Crunch dataset (deprecated)}
\usage{
\method{filter_}{CrunchDataset}(.data, ..., .dots)
}
\arguments{
\item{.data}{A \code{CrunchDataset}}

\item{...}{filter expressions}

\item{.dots}{More dots!}
}
\value{
\code{.data} with the filter expressions applied.
}
\description{
This function is deprecated, use \code{filter()} instead.
Applies a \code{CrunchLogicalExpression} filter to a
\code{CrunchDataset}. It's a "tidy" way of doing \code{ds[ds$var == val,]}.
}
