% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{GroupedCrunchDataset-class}
\alias{GroupedCrunchDataset-class}
\alias{GroupedCrunchDataset}
\title{A Crunch Dataset "Grouped By" Something}
\description{
This is a subclass of \code{crunch::CrunchDataset} that has a field for recording
"group_by" expressions.
}
\examples{
\dontrun{
ds <- loadDataset("Your dataset name")
class(ds) ## "CrunchDataset"
grouped_ds <- group_by(ds, var1)
class(grouped_ds) ## "GroupedCrunchDataset"
}
}
