% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-tibble.R
\name{is.selectedDimension}
\alias{is.selectedDimension}
\title{Check if a dimension is an MR selection dimension}
\usage{
is.selectedDimension(dims)
}
\arguments{
\item{dims}{A Crunch Cube Dimension}
}
\value{
A logical vector
}
\description{
Check if a dimension is an MR selection dimension
}
\keyword{internal}
