% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate-commands.R
\name{multiple_selection}
\alias{multiple_selection}
\title{Gather Variables into a Multiple Selection}
\usage{
multiple_selection(
  ...,
  labels = NULL,
  not_selected = NULL,
  exclude_empty = FALSE,
  hide_inputs = FALSE,
  title = NULL,
  description = NULL,
  notes = NULL
)
}
\arguments{
\item{...}{One or more crunch variables selected using their alias,
\code{\link[dplyr:across]{dplyr::across()}}, or the special selection functions in \code{\link{ca}}.}

\item{labels}{The labels for the subvariables (also called the subvariable
names in the crunch R package).}

\item{not_selected}{A vector of either strings or numbers indicating the category names
or codes that should be ignored. Note that R's \code{c()} function
will convert numbers to text if both are included, so if you want to specify both, use
the \code{list()} command instead.}

\item{exclude_empty}{Whether to exclude empty data.}

\item{hide_inputs}{A logical indicating whether to move the variables from collected
into the new array variable into the hidden folder}

\item{title}{Optional, the title of the variables created (also called the variable name
in the crunch R package)}

\item{description}{Optional, the description of the variables created}

\item{notes}{Optional, the notes of the variable created}
}
\description{
Create a Multiple Selection from one or more existing variables.
The new variable's alias is based on the name of the argument within
\code{mutate()}. Based on the Crunch Automation command
\href{https://help.crunch.io/hc/en-us/articles/360044079971}{CREATE MULTIPLE SELECTION}.
}
\details{
If \code{...} is specified using aliases or \code{dplyr::across()}, then
the other arguments can take functions based on the variables
selected using \code{\link{crplyr-formula-notation}}.
}
\seealso{
Other creating variables functions: 
\code{\link{categorical_array}()},
\code{\link{convert_to_text}()},
\code{\link{create-vars}},
\code{\link{multiple_dichotomy}()}
}
\concept{creating variables functions}
