% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate-automation.R
\name{mutate}
\alias{mutate}
\alias{mutate.CrunchDataset}
\title{Modify Variables in a Crunch Automation Script}
\usage{
\method{mutate}{CrunchDataset}(.data, ...)
}
\arguments{
\item{.data}{A crunch Dataset}

\item{...}{Arguments that create variables using the \code{\link{create-vars}} functions.}
}
\description{
Adds commands to a Crunch Automation script that modify the contents
of variables. Unlike \code{dplyr::mutate} on a traditional R \code{data.frame},
the variables are not created until you run the command \code{\link[=compute]{compute()}}.
}
\examples{
\dontrun{
ds <- ds \%>\%
    mutate(
       taste = categorical_array(
           taste_a, taste_b, 
           labels = c("Brand A", "Brand B"), 
           title = "Taste Rating"
       )
    )
    
# Show the query
show_query(ds)

# Execute the query 
ds <- ds \%>\% compute()
}
}
\seealso{
Other automation script commands: 
\code{\link{compute}()},
\code{\link{show_query}()}
}
\concept{automation script commands}
