% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automation-processing.R
\name{show_query}
\alias{show_query}
\alias{show_query.AutoReadyCrunchDataset}
\title{Show Text of Crunch Automation Query}
\usage{
\method{show_query}{AutoReadyCrunchDataset}(x, ...)
}
\arguments{
\item{x}{An \code{AutomationCrunchDataset} (a \code{CrunchDataset} modified with \code{\link[=mutate]{mutate()}} or similar)}

\item{...}{ignored}
}
\value{
invisibly, the text of the Crunch Automation Script
}
\description{
Show the text of the Crunch Automation script that you've prepared
using crplyr. To execute the query, see \code{\link[=compute]{compute()}}.
}
\examples{
\dontrun{
ds \%>\%
    mutate(
       taste = categorical_array(
           taste_a, taste_b, 
           labels = c("Brand A", "Brand B"), 
           title = "Taste Rating"
       )
    ) \%>\%
    show_query()
}
}
\seealso{
Other automation script commands: 
\code{\link{compute}()},
\code{\link{mutate}()}
}
\concept{automation script commands}
