% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{theme_crunch}
\alias{theme_crunch}
\title{Crunch ggplot theme}
\usage{
theme_crunch(base_size = 12, base_family = "sans")
}
\arguments{
\item{base_size}{Base text size}

\item{base_family}{Base text family}
}
\description{
Style ggplots according to Crunch style.
}
