% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{unweighted_n}
\alias{unweighted_n}
\title{Return the unweighted counts from summarize}
\usage{
unweighted_n()
}
\description{
This function allows you to return the unweighted counts from a Crunch dataset
or grouped crunch dataset. It can only be used from within a \code{summarise()}
call. If your dataset is unweighted, then unweighted_n() is equivalent to n().
}
\examples{
\dontrun{
ds \%>\%
   group_by(cyl) \%>\%
   summarize(
       raw_counts = unweighted_n(),
       mean = mean(wt)
   )
}
}
