% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-tibble.R
\name{as_tibble.CrunchCube}
\alias{as_tibble.CrunchCube}
\title{Flatten a Crunch Cube}
\usage{
\method{as_tibble}{CrunchCube}(x, return_real = FALSE, ...)
}
\arguments{
\item{x}{a CrunchCube}

\item{return_real}{Do you want to return the tibble represenatation of the
real cube with MR selection dimensions and missing values. Mostly useful
for debugging complex cubes.}

\item{...}{futher arguments passed on to \code{dplyr::as_tibble()}}
}
\description{
Crunch Cubes can be expressed as a long data frame instead of a
multidimensional array. In this form each dimension of the cube is a variable
of the dataframe and the cube values is expressed columns for each measure.
This is useful both to better understand what each entry of a cube
represents, and to work with the cube result using tidyverse tools.
}
