% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate}
\alias{mutate}
\alias{mutate.CrunchDataset}
\title{Mutate Crunch datasets (not implemented)}
\usage{
\method{mutate}{CrunchDataset}(.data, ...)
}
\arguments{
\item{.data}{A crunch Dataset}

\item{...}{Other arguments, currently ignored}
}
\description{
Just a method that returns a nicer error message. \code{mutate()} hasn't been
implemented yet. You can, however, derive expressions on the fly in
\code{\link[=summarize]{summarize()}}.
}
